requirements.txt
Flask==3.0.3
numpy
pyproj
SQLAlchemy
psycopg2-binary
geopandas==0.14.0
waitress
numpy
